+++ frontend/js/views/PreviewPage.tsx
@@ -1,5 +1,4 @@
 import React, {
-  SyntheticEvent,
   useCallback,
   useEffect,
   useMemo,
@@ -51,22 +50,44 @@ export default function PreviewPage(): JSX.Element {
     }
   }, [previewUrl]);
 
-  const onFrameNavigated = useCallback(
-    (event: SyntheticEvent<HTMLIFrameElement>) => {
-      const location = event.currentTarget.contentWindow?.location;
-      if (location && location.href !== previewUrl) {
-        if (location.href.startsWith(`${siteRootUrl}/`)) {
-          const url_path = location.href.substr(siteRootUrl.length);
-          get("/matchurl", { url_path }).then(({ exists, alt, path }) => {
-            if (exists) {
-              goToAdminPage("preview", path, alt);
-            }
-          }, showErrorDialog);
+  const onLoad = useCallback(() => {
+    const contentWindow = iframe.current?.contentWindow;
+    if (contentWindow) {
+      // Note that cross-origin security restrictions prevent us
+      // from being able to inspect or manipulate pages from
+      // other origins, so this will only work when the iframe is
+      // previewing one of our pages.
+      try {
+        // Pass keydown events on to parent window
+        // This is an attempt to ensure that hotkeys like Ctl-e work
+        // even when the iframe has the focus.
+        contentWindow.addEventListener("keydown", (ev: KeyboardEvent) => {
+          const clone = new KeyboardEvent(ev.type, ev);
+          window.dispatchEvent(clone) || ev.preventDefault();
+        });
+
+        const href = contentWindow.location?.href;
+        if (href && href !== previewUrl) {
+          // Iframe has been navigated to a new page (e.g. user clicked link)
+          if (href.startsWith(`${siteRootUrl}/`)) {
+            // Attempt to move Admin UI to new page
+            const url_path = href.substring(siteRootUrl.length);
+            get("/matchurl", { url_path }).then(({ exists, alt, path }) => {
+              if (exists) {
+                goToAdminPage("preview", path, alt);
+              }
+            }, showErrorDialog);
+          }
+        }
+      } catch (e) {
+        if (e instanceof DOMException && e.name === "SecurityError") {
+          // Ignore exceptions having to do with cross-origin restrictions
+        } else {
+          throw e;
         }
       }
-    },
-    [goToAdminPage, previewUrl, siteRootUrl]
-  );
+    }
+  }, [goToAdminPage, previewUrl, siteRootUrl]);
 
-  return <iframe className="preview" ref={iframe} onLoad={onFrameNavigated} />;
+  return <iframe className="preview" ref={iframe} onLoad={onLoad} />;
 }